# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evidrilo API",
    "version": "1.0.0",
    "description": "Public contract for the versioned Evidrilo API routes."
  },
  "paths": {
...
    "/v2/account/exports": {
      "post": {
        "operationId": "post_v2_account_exports",
        "tags": [
          "v2"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "202": {
            "description": "Successful response.",
            "content": {
              "application/vnd.evidrilo.account-export-job.v2+json": {
                "schema": {
                  "$ref": "#/components/schemas/account-export-job.v2"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/AccountExportJobEndpointTests.cs",
          "anchor": "/v2/account/exports"
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Stable key for replaying the same empty-body export creation request.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 80,
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          }
        ]
      },
      "parameters": []
    },
    "/v2/account/exports/{exportId}": {
      "get": {
        "operationId": "get_v2_account_exports_exportIdguid",
        "tags": [
          "v2"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/vnd.evidrilo.account-export-job.v2+json": {
                "schema": {
                  "$ref": "#/components/schemas/account-export-job.v2"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/AccountExportJobEndpointTests.cs",
          "anchor": "/v2/account/exports/{exportId:guid}"
        }
      },
      "parameters": [
        {
          "name": "exportId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "delete": {
        "operationId": "delete_v2_account_exports_exportIdguid",
        "tags": [
          "v2"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/vnd.evidrilo.account-export-job.v2+json": {
                "schema": {
                  "$ref": "#/components/schemas/account-export-job.v2"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/AccountExportJobEndpointTests.cs",
          "anchor": "/v2/account/exports/{exportId:guid}"
        }
      }
    },
    "/v2/account/exports/{exportId}/download": {
      "get": {
        "operationId": "get_v2_account_exports_exportIdguid_download",
        "tags": [
          "v2"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/vnd.evidrilo.account-export.v2+json": {
                "schema": {
                  "$ref": "#/components/schemas/account-export.v2"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/AccountExportJobEndpointTests.cs",
          "anchor": "/v2/account/exports/{exportId:guid}/download"
        }
      },
      "parameters": [
        {
          "name": "exportId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "RevenueCatWebhookAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      },
      "RevenueCatWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RevenueCat-Webhook-Signature"
      }
    },
    "schemas": {
...
      "account-export-job.v2": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/account-export-job.v2.json",
        "title": "Evidrilo account export job v2",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "exportId",
          "status",
          "requestId",
          "createdAt",
          "expiresAt"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.account-export-job"
          },
          "version": {
            "const": "2"
          },
          "exportId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "ready",
              "failed",
              "cancelled"
            ]
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{8,128}$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "artifactBytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 67108864
          },
          "errorCode": {
            "type": "string",
            "pattern": "^[A-Z0-9_]{1,64}$"
          }
        }
      },
...
      "account-export.v2": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/account-export.v2.json",
        "title": "Evidrilo complete account export v2",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "accountId",
          "generatedAt",
          "data",
          "requestId"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.account-export"
          },
          "version": {
            "const": "2"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "localDrafts",
              "profile",
              "attemptCommands",
              "syncChanges",
              "analyticsEvents",
              "recommendationEvents",
              "aiAuditEvents",
              "aiCreditConsent",
              "aiCreditGrants",
              "aiCreditReservations",
              "notificationPreferences",
              "entitlements",
              "entitlementEvents",
              "progressDaily",
              "studentProjects",
              "studentProjectRevisions",
              "studentProjectCloudConsent",
              "studentProjectCloudConsentEvents",
              "projectAiConsent",
              "projectAiConsentEvents",
              "projectAiActivity",
              "aiConversationSessions",
              "aiConversationTurnRequests"
            ],
            "properties": {
              "localDrafts": {
                "const": "not_on_server"
              },
              "profile": {
                "type": "object"
              },
              "attemptCommands": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "syncChanges": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "analyticsEvents": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "recommendationEvents": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "aiAuditEvents": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "requestId",
                    "promptVersion",
                    "provider",
                    "outcome",
                    "reasonCode",
                    "createdAt"
                  ],
                  "properties": {
                    "requestId": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9_-]{8,128}$"
                    },
                    "promptVersion": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80
                    },
                    "provider": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "outcome": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 64
                    },
                    "reasonCode": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "pattern": "^[A-Z0-9_]{1,64}$"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              },
              "aiCreditConsent": {
                "type": "object"
              },
              "aiCreditGrants": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "aiCreditReservations": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "notificationPreferences": {
                "type": "object"
              },
              "entitlements": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "entitlementEvents": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "progressDaily": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "studentProjects": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "studentProjectRevisions": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "studentProjectCloudConsent": {
                "$ref": "#/components/schemas/account-export.v1/properties/data/properties/studentProjectCloudConsent"
              },
              "studentProjectCloudConsentEvents": {
                "$ref": "#/components/schemas/account-export.v1/properties/data/properties/studentProjectCloudConsentEvents"
              },
              "projectAiConsent": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "policyVersion",
                  "granted",
                  "grantedAt",
                  "revokedAt",
                  "consentGeneration",
                  "updatedAt"
                ],
                "properties": {
                  "policyVersion": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "pattern": "^project-ai-data\\.v[0-9]+$"
                  },
                  "granted": {
                    "type": [
                      "boolean",
                      "null"
                    ]
                  },
                  "grantedAt": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time"
                  },
                  "revokedAt": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time"
                  },
                  "consentGeneration": {
                    "type": [
                      "integer",
                      "null"
                    ],
                    "minimum": 1
                  },
                  "updatedAt": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time"
                  }
                }
              },
              "projectAiConsentEvents": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "eventId",
                    "policyVersion",
                    "decision",
                    "consentGeneration",
                    "decidedAt"
                  ],
                  "properties": {
                    "eventId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "policyVersion": {
                      "type": "string",
                      "pattern": "^project-ai-data\\.v[0-9]+$"
                    },
                    "decision": {
                      "enum": [
                        "grant",
                        "revoke"
                      ]
                    },
                    "consentGeneration": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "decidedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              },
              "projectAiActivity": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "activityId",
                    "installationId",
                    "requestId",
                    "mode",
                    "projectId",
                    "stageId",
                    "operationId",
                    "baseProjectRevision",
                    "consentGeneration",
                    "resultProjectRevision",
                    "outcome",
                    "requestedSettlementOutcome",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "activityId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "installationId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "requestId": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9_-]{8,128}$"
                    },
                    "mode": {
                      "enum": [
                        "PROJECT",
                        "GENERAL"
                      ]
                    },
                    "projectId": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "uuid"
                    },
                    "stageId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "operationId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "baseProjectRevision": {
                      "type": [
                        "integer",
                        "null"
                      ],
                      "minimum": 1
                    },
                    "consentGeneration": {
                      "type": [
                        "integer",
                        "null"
                      ],
                      "minimum": 1
                    },
                    "resultProjectRevision": {
                      "type": [
                        "integer",
                        "null"
                      ],
                      "minimum": 1
                    },
                    "outcome": {
                      "enum": [
                        "PENDING",
                        "APPLIED",
                        "EDITED",
                        "DISMISSED",
                        "STALE",
                        "FAILED",
                        "COMPLETED"
                      ]
                    },
                    "requestedSettlementOutcome": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "APPLIED",
                        "EDITED",
                        "DISMISSED",
                        "STALE",
                        null
                      ]
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              },
              "aiConversationSessions": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "sessionId",
                    "creationRequestId",
                    "caseVersionId",
                    "turnCount",
                    "createdAt",
                    "updatedAt",
                    "expiresAt"
                  ],
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "creationRequestId": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9_-]{8,128}$"
                    },
                    "caseVersionId": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128
                    },
                    "turnCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 5
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              },
              "aiConversationTurnRequests": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "sessionId",
                    "requestId",
                    "turnIndex",
                    "status",
                    "startedAt",
                    "leaseExpiresAt",
                    "completedAt"
                  ],
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "requestId": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9_-]{8,128}$"
                    },
                    "turnIndex": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 5
                    },
                    "status": {
                      "enum": [
                        "reserved",
                        "completed",
                        "released"
                      ]
                    },
                    "startedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "leaseExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "completedAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    }
                  }
                }
              }
            },
            "not": {
              "anyOf": [
                {
                  "required": [
                    "prompts"
                  ]
                },
                {
                  "required": [
                    "responses"
                  ]
                },
                {
                  "required": [
                    "requestHash"
                  ]
                },
                {
                  "required": [
                    "settlementHash"
                  ]
                },
                {
                  "required": [
                    "activeRequestHash"
                  ]
                }
              ]
            }
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{8,128}$"
          }
        }
      },
...
      "http-errors.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/http-errors.v1.json",
        "title": "Evidrilo HTTP error response v1",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.http-error"
          },
          "version": {
            "const": "1"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          },
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "creditCost": {
            "description": "Present only when verified provider usage was charged even though the requested response was withheld.",
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        }
      },
...
    }
  }
}